[refactor] BoardDAO 에서 Service 로 일부 함수 추가 분리
